+++ example/test/test_services.dart
@@ -5,8 +5,8 @@ import 'dart:async';
 import 'dart:typed_data' as typed_data;
 
 import 'package:squadron/squadron.dart' as sq;
-import 'package:squadron/squadron.dart';
 
+import 'bigint_marshaler.dart' as marshaler;
 import 'generated/test_services.activator.g.dart';
 
 part 'generated/test_services.worker.g.dart';
@@ -256,26 +256,15 @@ class TestRequiredSuperParam {
   Future<void> clear() async {}
 }
 
-class _BigIntMarshaler implements GenericMarshaler<BigInt> {
-  // "const" so it can be used to annotate parameters and return values
-  const _BigIntMarshaler();
-
-  @override
-  dynamic marshal(BigInt data) => data.toString();
-
-  @override
-  BigInt unmarshal(dynamic data) => BigInt.parse(data);
-}
-
 @sq.SquadronService()
 class TestBigInt {
   TestBigInt();
 
   @sq.squadronMethod
-  @_BigIntMarshaler()
+  @marshaler.BigIntMarshaler()
   FutureOr<BigInt> add(
-    @_BigIntMarshaler() BigInt a,
-    @_BigIntMarshaler() BigInt b,
+    @marshaler.BigIntMarshaler() BigInt a,
+    @marshaler.BigIntMarshaler() BigInt b,
   ) =>
       a + b;
 }