+++ lib/src/assets/worker_assets_worker.dart
@@ -2,9 +2,8 @@ part of 'worker_assets.dart';
 
 extension on WorkerAssets {
   /// Worker
-  String _generateWorker(
-      List<SquadronMethodReader> commands, List<DartMethodReader> unimplemented,
-      {bool finalizable = false}) {
+  String _generateWorker(List<SquadronMethodReader> commands,
+      List<DartMethodReader> unimplemented, bool finalizable) {
     final serialized = _service.parameters.serialize();
     var activationArgs = serialized.isEmpty
         ? '$_serviceActivator($TSquadron.platformType)'
@@ -48,7 +47,7 @@ extension on WorkerAssets {
 
           ${unimplemented.map((m) => m.unimplemented()).join('\n\n')}
 
-          ${_service.accessors.map((a) => a.unimplemented(_typeManager)).join('\n\n')}
+          ${_service.accessors.where((a) => !a.isOperationMap).map((a) => a.unimplemented(_typeManager)).join('\n\n')}
 
           ${finalizable ? 'final $TObject _detachToken = $TObject();' : ''}
         }
@@ -98,7 +97,7 @@ extension on WorkerAssets {
 
           ${unimplemented.map((cmd) => cmd.forwardTo(instance)).join('\n\n')}
 
-          ${_service.accessors.map((a) => a.forwardTo(instance, _typeManager)).join('\n\n')}
+          ${_service.accessors.where((a) => !a.isOperationMap).map((a) => a.forwardTo(instance, _typeManager)).join('\n\n')}
 
           ${_typeManager.getWorkerOverrides().entries.map((e) => _forwardOverride(e.key, instance, e.value)).join('\n\n')}