+++ lib/src/marshalers/marshaler_explicit.dart
@@ -1,20 +1,23 @@
 part of 'marshaler.dart';
 
 class _ExplicitMarshaler extends Marshaler {
-  _ExplicitMarshaler(this._marshaler, this._marshalerType) {
+  _ExplicitMarshaler(
+      TypeManager typeManager, this._marshaler, this._marshalerType) {
     _itemType = _marshalerType.typeArguments.first;
+    _pivotType = (_marshalerType.typeArguments.length > 1)
+        ? _marshalerType.typeArguments.last
+        : null;
     final variable = _marshaler.variable;
     if (variable != null) {
-      if (variable.enclosingElement is ClassElement) {
+      if (variable.enclosingElement is InterfaceElement) {
         _instance = '${variable.enclosingElement!.name!}.${variable.name}';
       } else {
         _instance = variable.name;
       }
     } else {
-      var typeName = (_marshaler.toTypeValue() ?? _marshaler.type).toString();
-      if (typeName.endsWith('?') || typeName.endsWith('*')) {
-        typeName = typeName.substring(0, typeName.length - 1);
-      }
+      final typeName = typeManager
+          .handleDartType(_marshaler.toTypeValue() ?? _marshaler.type!)
+          .getTypeName(NullabilitySuffix.none);
       _instance = '(const $typeName())';
     }
   }
@@ -23,15 +26,20 @@ class _ExplicitMarshaler extends Marshaler {
   final ManagedType _marshalerType;
   late final String _instance;
   late final ManagedType _itemType;
+  late final ManagedType? _pivotType;
 
   @override
   bool targets(ManagedType type) => type.dartType?.isA(_itemType) ?? false;
 
   @override
   String serializerOf(ManagedType type, Converters converters) =>
-      '((\$) => $_instance.marshal(\$))';
+      (_pivotType?.dartType is DynamicType)
+          ? '$_instance.marshal'
+          : '((\$) => $_instance.marshal(\$))';
 
   @override
   String deserializerOf(ManagedType type, Converters converters) =>
-      '((\$) => $_instance.unmarshal(\$))';
+      (_pivotType?.dartType is DynamicType)
+          ? '$_instance.unmarshal'
+          : '((\$) => $_instance.unmarshal(\$))';
 }